src/dev_assistant/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.11.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
anthropic==0.42.0
python-multipart==0.0.20
gitpython==3.1.44
numpy==2.2.1
anyio>=4.5.0
